Move AST to package
